changes metric data type

+++ app/utils/constants.py
@@ -282,9 +282,9 @@ DATA_TYPES: Dict[str, Callable[[str], Any]] = {
     "weeklyPeopleVaccinatedSecondDoseByVaccinationDate": int,
     "cumPeopleVaccinatedSecondDoseByVaccinationDate": int,
 
-    "cumVaccinationFirstDoseUptakeByPublishDatePercentage": int,
-    "cumVaccinationSecondDoseUptakeByPublishDatePercentage": int,
-    "cumVaccinationCompleteCoverageByPublishDatePercentage": int,
+    "cumVaccinationFirstDoseUptakeByPublishDatePercentage": float,
+    "cumVaccinationSecondDoseUptakeByPublishDatePercentage": float,
+    "cumVaccinationCompleteCoverageByPublishDatePercentage": float,
 }
 
 # Values must be provided in lowercase characters.
@@ -551,7 +551,7 @@ if ENVIRONMENT == "DEVELOPMENT":
         'newCasesPCROnlyBySpecimenDate': int,
         'newCasesLFDConfirmedPCRBySpecimenDateRollingSum': int,
 
-        "cumVaccinationFirstDoseUptakeByPublishDatePercentage": int,
-        "cumVaccinationSecondDoseUptakeByPublishDatePercentage": int,
-        "cumVaccinationCompleteCoverageByPublishDatePercentage": int,
+        "cumVaccinationFirstDoseUptakeByPublishDatePercentage": float,
+        "cumVaccinationSecondDoseUptakeByPublishDatePercentage": float,
+        "cumVaccinationCompleteCoverageByPublishDatePercentage": float,
     }
